refactor: request 폴더 controller 밑으로 이동 #236
