+++ src/main/java/com/ceos/bankids/controller/ChallengeController.java
@@ -11,8 +11,8 @@ import com.ceos.bankids.dto.KidWeekDTO;
 import com.ceos.bankids.dto.ProgressDTO;
 import com.ceos.bankids.dto.WeekDTO;
 import com.ceos.bankids.mapper.ChallengeMapper;
-import com.ceos.bankids.mapper.request.ChallengeRequest;
-import com.ceos.bankids.mapper.request.KidChallengeRequest;
+import com.ceos.bankids.controller.request.ChallengeRequest;
+import com.ceos.bankids.controller.request.KidChallengeRequest;
 import io.swagger.annotations.ApiOperation;
 import java.util.List;
 import javax.validation.Valid;
