+++ src/main/java/com/ceos/bankids/controller/request/KakaoRequest.java
@@ -1,4 +1,4 @@
-package com.ceos.bankids.mapper.request;
+package com.ceos.bankids.controller.request;
 
 import io.swagger.annotations.ApiModelProperty;
 import javax.validation.constraints.NotNull;
@@ -13,9 +13,9 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class FamilyRequest {
+public class KakaoRequest {
 
-    @ApiModelProperty(example = "863035e8-b067-4dae-8961-b52e6839a81b")
+    @ApiModelProperty(example = "erYw3skzA27NEcbri4hIu18MZ7PRKzhsvQB31JU-iyq1BWgzYv1xve8KpjV_24CwGx3PiwopyNgAAAGBxHM1jQ")
     @NotNull(message = "code may not be null")
     private String code;
 }
