+++ src/main/java/com/ceos/bankids/controller/request/WithdrawalRequest.java
@@ -1,4 +1,4 @@
-package com.ceos.bankids.mapper.request;
+package com.ceos.bankids.controller.request;
 
 import io.swagger.annotations.ApiModelProperty;
 import javax.validation.constraints.NotNull;
@@ -13,9 +13,10 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class KakaoRequest {
+public class WithdrawalRequest {
 
-    @ApiModelProperty(example = "erYw3skzA27NEcbri4hIu18MZ7PRKzhsvQB31JU-iyq1BWgzYv1xve8KpjV_24CwGx3PiwopyNgAAAGBxHM1jQ")
-    @NotNull(message = "code may not be null")
-    private String code;
+    @ApiModelProperty(example = "나 탈퇴하겠어!")
+    @NotNull(message = "message may not be null")
+    private String message;
+
 }
