+++ src/main/java/com/ceos/bankids/service/ChallengeServiceImpl.java
@@ -2,7 +2,7 @@ package com.ceos.bankids.service;
 
 import com.ceos.bankids.constant.ChallengeStatus;
 import com.ceos.bankids.constant.ErrorCode;
-import com.ceos.bankids.mapper.request.KidChallengeRequest;
+import com.ceos.bankids.controller.request.KidChallengeRequest;
 import com.ceos.bankids.domain.Challenge;
 import com.ceos.bankids.domain.ChallengeCategory;
 import com.ceos.bankids.domain.ChallengeUser;
