+++ src/main/java/com/ceos/bankids/service/KakaoService.java
@@ -1,6 +1,6 @@
 package com.ceos.bankids.service;
 
-import com.ceos.bankids.mapper.request.KakaoRequest;
+import com.ceos.bankids.controller.request.KakaoRequest;
 import com.ceos.bankids.dto.oauth.KakaoTokenDTO;
 import com.ceos.bankids.dto.oauth.KakaoUserDTO;
 import org.springframework.stereotype.Service;
