+++ src/main/java/com/ceos/bankids/service/KakaoServiceImpl.java
@@ -1,7 +1,7 @@
 package com.ceos.bankids.service;
 
 import com.ceos.bankids.constant.ErrorCode;
-import com.ceos.bankids.mapper.request.KakaoRequest;
+import com.ceos.bankids.controller.request.KakaoRequest;
 import com.ceos.bankids.dto.oauth.KakaoTokenDTO;
 import com.ceos.bankids.dto.oauth.KakaoUserDTO;
 import com.ceos.bankids.exception.BadRequestException;
