+++ src/main/java/com/ceos/bankids/service/UserServiceImpl.java
@@ -3,8 +3,8 @@ package com.ceos.bankids.service;
 import com.ceos.bankids.domain.User;
 import com.ceos.bankids.dto.OptInDTO;
 import com.ceos.bankids.dto.UserDTO;
-import com.ceos.bankids.mapper.request.ExpoRequest;
-import com.ceos.bankids.mapper.request.UserTypeRequest;
+import com.ceos.bankids.controller.request.ExpoRequest;
+import com.ceos.bankids.controller.request.UserTypeRequest;
 import com.ceos.bankids.repository.UserRepository;
 import java.util.List;
 import java.util.Optional;
