+++ src/test/java/com/ceos/bankids/unit/controller/FamilyControllerTest.java
@@ -13,7 +13,7 @@
 //import com.ceos.bankids.exception.ForbiddenException;
 //import com.ceos.bankids.mapper.FamilyController;
 //import com.ceos.bankids.mapper.NotificationMapper;
-//import com.ceos.bankids.mapper.request.FamilyRequest;
+//import FamilyRequest;
 //import com.ceos.bankids.repository.FamilyRepository;
 //import com.ceos.bankids.repository.FamilyUserRepository;
 //import com.ceos.bankids.service.ChallengeServiceImpl;
